sphinx: mock imports in conf.py
Rather than doing it in the code, detect and mock in conf.py.
